Move Image and Text styles into a separate file
